Mirotor/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
Django==4.2
django-filter==23.1
django-request==1.6.2
djangorestframework==3.14.0
fonttools==4.39.3
frozenlist==1.3.3
gunicorn==20.1.0
idna==3.4
kiwisolver==1.4.4
magic-filter==1.0.9
matplotlib==3.7.1
multidict==6.0.4
numpy==1.24.2
packaging==23.1
Pillow==9.5.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==1.0.0
psycopg2-binary==2.9.6
pytz==2023.3
requests==2.28.2
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.3
urllib3==1.26.15
yarl==1.8.2
